+++ src/app/components/coaches/coaches.component.scss
@@ -7,7 +7,7 @@
 }
 
 .example-header-image {
-  background-image: url('../../assets/img/isa-logo.png');
+  background-image: url('../../../assets/img/isa-logo.png');
   background-size: cover;
   background-repeat: no-repeat;
   background-position: center;
