+++ src/app/components/header/header.component.ts
@@ -3,7 +3,7 @@ import {MatToolbarModule} from '@angular/material/toolbar';
 import {MatIconModule} from '@angular/material/icon';
 import {MatMenuModule} from '@angular/material/menu';
 import { Router } from '@angular/router';
-import { RoutesConstants } from '../shared/utils/constants';
+import { RoutesConstants } from '../../shared/constants/routes.constant';
 import { MatButtonModule } from '@angular/material/button';
 
 @Component({
