+++ src/app/components/home/home.component.scss
@@ -3,7 +3,7 @@
 }
 
 .example-header-image {
-  background-image: url('../../assets/img/isa-logo.png');
+  background-image: url('../../../assets/img/isa-logo.png');
   background-size: cover;
   background-repeat: no-repeat;
   background-position: center;
