+++ src/app/components/players/player/player.component.html
