+++ src/app/components/players/players.component.ts
