+++ src/app/services/player.service.spec.ts
@@ -1,0 +1,23 @@
+// import { TestBed } from '@angular/core/testing';
+// import { PlayerService } from './player.service';
+
+// describe('PlayerService', () => {
+//   let service: PlayerService;
+
+//   beforeEach(() => {
+//     TestBed.configureTestingModule({});
+//     service = TestBed.inject(PlayerService);
+//   });
+
+//   it('should be created', () => {
+//     expect(service).toBeTruthy();
+//   });
+
+//   it('should initialize with default values', () => {
+//     expect(service).toBeDefined();
+//   });
+
+//   it('should handle player data', () => {
+//     expect(service).toBeTruthy();
+//   });
+// });
