+++ src/app/services/player.service.ts
@@ -1,0 +1,40 @@
+import { inject, Injectable } from '@angular/core';
+import { HttpClient } from '@angular/common/http';
+import { Observable } from 'rxjs';
+
+export interface Player {
+  id: string;
+  name: string;
+  email: string;
+  level: number;
+  score: number;
+}
+
+@Injectable({
+  providedIn: 'root'
+})
+export class PlayerService {
+  private apiUrl = '/api/players';
+
+  private http = inject(HttpClient);
+
+  getPlayers(): Observable<Player[]> {
+    return this.http.get<Player[]>(this.apiUrl);
+  }
+
+  getPlayer(id: string): Observable<Player> {
+    return this.http.get<Player>(`${this.apiUrl}/${id}`);
+  }
+
+  createPlayer(player: Player): Observable<Player> {
+    return this.http.post<Player>(this.apiUrl, player);
+  }
+
+  updatePlayer(id: string, player: Player): Observable<Player> {
+    return this.http.put<Player>(`${this.apiUrl}/${id}`, player);
+  }
+
+  deletePlayer(id: string): Observable<void> {
+    return this.http.delete<void>(`${this.apiUrl}/${id}`);
+  }
+}
